Refac: mvc pattern
