+++ db_connection_mvc/controllers/employee_controller.py
@@ -1,25 +1,34 @@
+from fastapi import APIRouter
 from services.employee_service import EmployeeService
-from typing import List, Dict, Any
+
+router = APIRouter()
+employee_service = EmployeeService()
 
 
-class EmployeeController:
-    def __init__(self):
-        self.employee_service = EmployeeService()
+# 유저 목록 조회
+@router.get("/employees")
+async def get_all_employees():
+    return await employee_service.get_all_employees()
 
-    def format_employee_list(self, employees: List[Dict[str, Any]]) -> Dict[str, Any]:
-        return {
-            "status": "success",
-            "data": employees,
-            "message": "직원 목록을 성공적으로 조회했습니다.",
-            "count": len(employees),
-        }
 
-    def format_error(self, error_message: str) -> Dict[str, Any]:
-        return {"status": "error", "data": None, "message": error_message, "count": 0}
+# class EmployeeController:
+#     def __init__(self):
+#         self.employee_service = EmployeeService()
 
-    async def get_all_employees(self) -> Dict[str, Any]:
-        try:
-            employees = self.employee_service.get_all_employees()
-            return self.format_employee_list(employees)
-        except Exception as e:
-            return self.format_error(str(e))
+#     def format_employee_list(self, employees: List[Dict[str, Any]]) -> Dict[str, Any]:
+#         return {
+#             "status": "success",
+#             "data": employees,
+#             "message": "직원 목록을 성공적으로 조회했습니다.",
+#             "count": len(employees),
+#         }
+
+#     def format_error(self, error_message: str) -> Dict[str, Any]:
+#         return {"status": "error", "data": None, "message": error_message, "count": 0}
+
+#     async def get_all_employees(self) -> Dict[str, Any]:
+#         try:
+#             employees = self.employee_service.get_all_employees()
+#             return self.format_employee_list(employees)
+#         except Exception as e:
+#             return self.format_error(str(e))
