--- db_connection_mvc/views/__init__.py
@@ -1,3 +1,0 @@
-from .employee_view import router
-
-__all__ = ["router"]
